+++ R/semicolon_terminator_linter.R
@@ -5,6 +5,7 @@
 #' }
 #' @export
 semicolon_terminator_linter <- function(semicolon = c("compound", "trailing")) {
+  semicolon <- match.arg(semicolon, several.ok = TRUE)
   function(source_file) {
     tokens <- with_id(source_file, ids_with_token(source_file, "';'"))
     is_trailing <- is_trailing_sc(tokens, source_file)